Narrowed the setting scope to include thisProjectRef
thisProjectRef was necessary to narrow the setting test to the project associated with it in order to prevent other scopes stomping over the value.

The setting logic has also been simplified by using a conditional task assignment.

The Visualizer was tested to work in the ConductR project i.e. bundle:dist, conduct load, conduct run...

The Visualizer also gets bundled correctly for the ConductR packaging i.e. conductr/docker:publishLocal and then unpack the resulting samples/visualizer package of the docker stage folder. In addition I ran the sandbox against the new image and used the --with-features visualizer.